🔨 refactor(types): 공통 타입 정의 통합으로 일관된 타입 관리 구조 개선
- src/advanced/types/index.ts 생성하여 타입 정의 중앙화
- Product, CartItem, DiscountInfo, DiscountResult, PointsResult 등 핵심 타입 통합
- CartItemProps, ShoppingGuideProps, UseDiscountTimersProps 등 컴포넌트 Props 타입 정의 이동
- 각 파일에서 중앙 타입 모듈을 import하도록 수정
- 타입 중복 제거 및 일관성 있는 관리 구조로 유지보수성 향상
