+++ src/advanced/components/guide/ShoppingGuide.tsx
@@ -1,6 +1,4 @@
-interface ShoppingGuideProps {
-  onClose: () => void;
-}
+import { ShoppingGuideProps } from '../../types';
 
 const ShoppingGuide = ({ onClose }: ShoppingGuideProps) => {
   const handleOverlayClick = (e: React.MouseEvent) => {
